requirements.txt
spotipy
python-dotenv
requests
pyTelegramBotAPI
